Fixing wui after changes in model, everything compiles but needs testing

+++ eppic-model/src/main/java/eppic/model/ResidueBurialDB.java
@@ -45,7 +45,7 @@ public class ResidueBurialDB implements Serializable {
 		this.bsa = bsa;
 	}
 	
-	public int getRegion() {
+	public short getRegion() {
 		return this.region;
 	}
